Support a site-wide configuration script for ignored extensions.  I know, it's my time I'm wasting by not consigning these non-source files to a non-source sub-directory.  But it's taking me ages to get around to beating part of BlueArc's build system back into compliance with that rule (after someone else broke it) and searching the generated non-source files in question takes long enough to cause a logged warning.
src/e/edit/EErrorsWindow.java: Defer to the Workspace(FileList) for our FileIgnorer, so it gets to worry about when FileIgnorer's configuration should be cached and when it should be reassessed.

src/e/edit/WorkspaceFileList.java: We should definitely reassess the FileIgnorer configuration when the user asks us to rescan and it's cheap enough to reassess any time we decide to rescan ourselves.  It might be nice to notice that the script that's at the head of the PATH has changed but that's more effort than it's worth.  Anyway, the script is *not* cheap enough to reassess for each file that we index...

src/e/edit/FileIgnorer.java: ... which was why isIgnoredExtension used to cache once, on the first call.  That caching meant that you had to restart Evergreen to get it to notice changes to the "Don't index" field of the preferences dialog.  Now I'm adding another way to configure the ignored extensions list - another local script - so it becomes more important to get the caching right.

What to call the new script?  "echo" and "local" seem non-controversially OK.  Elliott wants to mention not-indexing instead of just "uninteresting" (in which context?) or "non-source" (they might be generated source that happens never to be interesting).  I wanted to mention Evergreen, or to give some other indication that we mean "don't index in your programmer's editor".

Elliott's egging me on to change echo-local-non-source-directory-pattern to "echo-local-directories-evergreen-should-not-index".  The BlueArc misc repository is currently locked, so that would be inconvenient today.  But I do like the pluralization and I'm happy with relegating the fact that it's interpreted as a list of patterns to the documentation.

We're also talking about removing the preference of default extensions, in favor of hard-coding the list in the source.  I imagine we'd just remove the documentation and the UI but keep the code for reading any existing preference.

www/index.html: Sketchily document the new feature and a not-widely-known and probably never used feature of the existing site-wide script (the way you can use `pwd` to produce per-workspace configuration).  This wants a more invasive edit, perhaps tomorrow, when I rename the directory pattern script.