lib/build/universal.make: Write quite a bit of extra code so we can do without the comment about the harmlessness of the duplication when we run findbugs in salma-hayek.  Remove the $(addprefix) complexity in the findbugs rule, and where it was copied from.  I presume the original reason was that the argument might be empty.  It's not now and doesn't seem likely to be again.
The extra colon on the end of auxclasspath happens when EXTRA_JARS ends in a space, because the last wildcard to be appended evaluates empty.  makeNativePath can strip trailing spaces and solve that more widely.  Evaluate the wildcard immediately, as usual, so we minimize the file system chatter.

lib/build/filter-build-output.rb: I was going to ask what findbugs's output looks like, thinking that if it doesn't generate any warnings, we wouldn't want to see it.  Well, there doesn't seem much danger of it not generating any warnings.  They look easy enough to spot, especially the warning counter at the end.  There are several other approaches we could use to solve this sort of problem but, while we're just generating things that really do look like compiler warnings, I think I'll stick to the existing mechanism.  exit 1 was, though, a good (clean, simple) hack.  I'm tempted to change the echo.% rule to work that way too.