src/e/edit/FileIgnorer.java: src/e/edit/WorkspaceFileList.java: add a few intention-revealing "final"s, make it clear that getIgnoredExtensions wasn't a getter by changing its name, make it clear that it doesn't modify the ignoredExtensions field by using Beck's idiomatic name "result", avoid calling File.getName twice on the same File (for a pathetic saving), and avoid calling File.isDirectory twice on the same File to save a poor but worthwhile 10% of the indexing time on a large (30,000 file) in-cache tree.
Also add a really important comment to nameEndsWithOneOf to stop me fiddling with that method now I've actually run the right test and found that it wouldn't make any difference if I worked out how to make it free.

At the moment, most of the cost of indexing turns out to be due to File.isHidden and -- to a much greater extent -- FileUtilities.isSymbolicLink. The former we could trivially lose without affecting Unix (the implementation is appallingly expensive, because it assumes we're on Windows and need to actually call getBooleanAttributes rather than just examine the name), but I don't know whether removing it would cause problems for Windows (though presumably the cost is even worse on Windows, where everything related to the file system is slow). As for testing for symbolic links, we could either write our own JNI (which would help with the other stuff too, since we could implement something more like stat(2) instead of asking for the same information one bit at a time), or wait for JDK7's nio improvements.