lib/build/file-list-to-wxi.rb: Shortcuts and shell extensions are seemingly typically created immediately after the File to which they refer and must, in some cases, be inside the same Component.  I used this hook several times this evening in various experiments and I expect I'll use it again.
lib/build/installer.wxs: Skeletal support for an open-here feature.  I tried a couple of ways of filling in this Component this evening.  I suspect I'll be back.

lib/build/universal.make: Open-here requires another random number.

You know, all this, and the non-maintainability of the HKLM and HKCU variants of the code in that .wxs file make me think that I should have one Ruby script that generates the whole thing.  I suspect it'd be more readable and it'd certainly be better commented.

I'm also about to check-in the Terminator TODO file with some related .wxs snippets for revisiting on another evening...