lib/build/file-list-to-wxi.rb: Here's the working code to create an advertised shortcut... but effectively commented-out because...
lib/build/installer.wxs: ... the thread at the URL here suggests that an advertised shortcut can't be made optional.  That plus the desire to let end users tinker with the shortcuts is sufficient to make me lean towards keeping the gruesome HKCU hack which, as well as being the code we've used for a long time, is also the suggestion at the end of that thread.