src/e/util/InAppServer.java: On Linux, Ruby seems able to connect(2) to 0.0.0.0, which was what we were writing into edit-server-port.  On Windows, it throws EADDRNOTAVAIL, perhaps not unreasonably.  The intention of this code, in the case where we're binding to INADDR_ANY, was to write the host name into the file, so you could run bin/evergreen from another machine and have it throw the file into the Evergreen running on your primary desktop.  That obviously hasn't been working on any platform for quite some time.
This code has two callers and only one passes in null.  We could move this code into that caller.  I don't want to do that because INADDR_ANY would be a better address to listen on for a multi-homed host, even if we don't choose to write all of its addresses into the server-port file.