bin/invoke-java.rb: The code to act on RUBY_LAUNCHER_INVOKING can never have worked.  The only reason it can have given the impression of working is that in the in-app client hasn't been working for Windows for ages.
bin/show-alert.rb: If I put it here, then it does work, although this code isn't anything to do with reporting exceptions.  "launch" was a better name for its home.  "launch" and "launch0" give the impression that the intention was for eg Terminator to sub-class invoke-java.rb's Java class.  Perhaps that would be the cleaner solution, though I'm not a fan of derivation except for interface implementation.