LC Spencer demonstrates that the error we throw when Ruby isn't found needs to be as easily understood as it once was and as our error when we can't find Cygwin again seems to be.
native/all/ruby-launcher/ruby-launcher.cpp: We only invoke this program in one place, where we give it the full path to rubyw, so there's no point in causing the C runtime the overhead of execvp over execv, and no point in muddying the waters of the error message with the suggestion that we might need to do a path search.

native/Mingw/cygwin-launcher/cygwin-launcher.cpp: Add a couple of .exe extensions, just in case execv without the p doesn't and so that we can use checkReadableFile to throw an error if Cygwin Ruby is missing.  Move the verbiage that details our Cygwin and Cygwin Ruby requirements higher up, so that it will be output for any exception, not just ones encountered trying to find Cygwin's bin directory.

I haven't tried to compile this, let alone run it or test it in any way.