lib/build/per-directory.make: Reapply r2637, minus the FIXME...
lib/build/universal.make: ... for which a simple solution occurs today.  This should also let Elliott build salma-hayek from clean.  I can see how that might have been broken previously, though it was working for me by (bad) luck.

My goal with the per-directory.make change was to tell make to compile java-launcher.cpp, if it's been changed, before using it via javahpp.  That missing dependency became apparent with the addition of JNI to salma-hayek, when I was investigating our behavior on Windows when Java crashes, as it was doing repeatedly for Ron Pagani the other day.