bin/ithaki.sh: Pay for the download only when we're making the Packages file that cares about the md5sums.  It took some fighting with various rsync syntaxes but it turns out not to take much code, or time, or bandwidth.
lib/build/universal.make: That should be enough as long as we don't overwrite the build on the server when the version number hasn't changed.  I've perhaps shortened the message too much.  I hope that won't lead to anyone contriving to upload a .deb with the same version number as one that's already in a Packages file.

If you want to change the message, note that filter-build-output.rb requires these progress messages to be of the form "Verbing blah...", so you can't just stick in a "Not ", at least not without complicating that script too.

As ever, with nightly build changes, this is only partly tested.