bin/show-alert.rb: When Ruby throws a Windows MessageBox, the same hint about Ctrl-C applies as when the C++ throws one.  Let's display the same message.  When we report an unhandled exception, all of our callers seem to assume that we'll exit with failure, so let's do that.  It's important in, for example, the case that Java crashed under javahpp, so make knows to stop...
lib/build/per-directory.make: ... although make not knowing the right place to start again pointed out a bug here.  Never write "> $@" in a rule, even when you have to write it in a strange way to accommodate the local JNI_CLASS_NAME variable.  Hmm... and that cp's not atomic either.  Tsk tsk.  I'll be back here.

bin/invoke-java.rb: Throw an exception, and hence a MessageBox (on similar) if the Java program we run fails, for example because Java or the JNI crashes.  I've seen two crashes in as many weeks.  Ron Pagani's was particularly nasty in that it looked like a Terminator installation problem.  He double-clicked the desktop shortcut and "nothing happened".