lib/build/per-directory.make: Fix the atomicity of the cp of NEW_JNI_HEADER to JNI_HEADER.  It seems possible to write this javahpp rule in a much more conventional way, as two rules, with liberal (but atomic!) use of $@.  The reason it was written in such an unusual way was that the value of JNI_CLASS_NAME won't be the same by the time the rule is run.  It seems that we can cache that in a target-specific variable in an intention-revealing way.
I've done a few test builds from clean and from already compiled, in both Salma Hayek and Terminator, paying particular attention to the case where I've touched the Java.  There, I checked that we're still regenerating the JNI header when the Java changes but that we're not recompiling the corresponding C++ unless there's been any content change in the JNI header.  It all seems to work.  Bon, it certainly looks easier to maintain.