native/all/pty/PtyGenerator.h: Report an error message, even though it's not one that would mean something to most users, if we can't open the slave end of the pty.  An easy way to do that seems to be checking it in the parent before forking.  Cygwin has a hard-coded limit (via an array, iirc) on the number of ptys.  I think today's report, from Ron Pagani, was the second time we've seen someone run out.  In Ron's case, a previous Terminator crash looks to have left him with lots of orphaned bash processes.  I can say that with some confidence because the same happened to me when I first tried to reproduce his issue with 64 terminals.
In other bad but probably unrelated news today, Mike Nguyen reports that his 10 tab Terminator running on XP always hangs by the time he comes back in the morning.  He was very keen on the idea that it was the log file size that was doing it.  I think it was more likely that he'd run out of memory.  His new XP machine has 2 GiB of RAM, so I've relaxed the "100m" limit in his invoke-java.rb back to the default "1g".  If that does "fix" it, then it'll be more evidence that it's worth trying to divine the amount of RAM... or bounding the scrollback.