src/e/edit/TagsUpdater.java: Avoid inspecting the text area from the background, unsafely.  Instead, pass it to...
src/e/edit/TagsPanel.java: ... the renderer, which is presumably running on the EDT.

src/e/edit/TagReader.java: The TagReader no longer needs to cache the toolTip.

If anything, this seems to have made the TagsUpdater stopwatch take longer on the stone1/.../soapC.cpp test case, but not outside the bounds of the possible noise.  In any case, safety was the motivator here.  Elliott said "the only reason i can think of for doing the text-getting on the wrong thread is that we were perhaps worried about the text changing without the tags updating. but since we re-tag whenever a newline is added or removed, that seems like a small window. and the current code, however well-intentioned, is unsafe (even though afaik, we've never had any trouble from it)".  He's forgetting, I think, that it wasn't unsafe, I believe, until I removed the publish->process split, at his suggestion, just yesterday.