lib/build/universal.make: Always use absolute filenames with the build-finished target...
lib/build/per-directory.make: ... so we have a rule for building $(SALMA_HAYEK)/.generated/java.build-finished when building salma-hayek from clean, so that make -j builds from clean.

This was what the change backed out in my previous check-in was intended to fix, not plain old building from clean, as I tested and claimed in my check-in comment, but parallel building from clean.