src/e/util/StringHistory.java: fix a bug equivalent to that fixed by mad in Evergreen revision 1461...
src/e/util/StringUtilities.java: ...by moving the responsibility here. It's only incidental that we no longer need to create the big intermediate String (though that may be handy someday), but it is deliberate that we still hard-code the use of "\n" rather than switching to the more obvious PrintWriter.println; we could switch, but I'm not sure we want to, and even if we do, we should make sure that readFile copes with non-Unix line endings first.