native/Mingw/cygwin-launcher/cygwin-launcher.cpp: An HKEY_CURRENT_USER installation is more specific than an HKEY_LOCAL_MACHINE one so, thinking about this for what the history suggests is the first time, I think we should prefer per-user installations.  Our error message here would be improved by saying "HKEY_LOCAL_MACHINE" instead of its pointer value...
native/Headers/HKEY.h: ... which wasn't previously sufficient reason for me to both with this but...

native/all/java-launcher/java-launcher.cpp: ... now I want to convert the value to a Cygwin /proc/registry path.  I supposed I could have used strings but I wanted to copy the hive iteration from cygwin-launcher.cpp verbatim.  I became a little fed up of the duplication of the exception handling, so I collapsed one copy at the same time as adding another.  (It was that or nest loops.)  The "Jesse Kriss" comment seemed to have drifted away from the code it referred to, so I fixed that too.

This may fix Ken Harris's problem, as he seems to have a per-user installation of Java.  I wonder why we haven't seen that before.

I haven't even compiled this on Windows yet.