src/e/util/StringUtilities.java: We have three calls to Pattern.quote and rather more to regularExpressionFromLiteral, whose difference from Pattern.quote is oddly not documented.  We don't seem to have any code to quote replacement strings, yet, as the JavaDoc for Matcher says [
Note that backslashes (\) and dollar signs ($) in the replacement string may cause the results to be different than if it were being treated as a literal replacement string. Dollar signs may be treated as references to captured subsequences as described above, and backslashes are used to escape literal characters in the replacement string.

].  Perhaps this new replacementStringFromLiteral, and regularExpressionFromLiteral, should move to PatternUtilities.  That was the first place I looked, though there was no way I was giving up until I'd looked here too.

The motivating call-site for replacementStringFromLiteral is in Evergreen and so in a different check-in.