src/e/edit/Evergreen.java: Reinstate Elliott's desired and reasonable wish for the behavior of Close Workspace, adding a new action that's...
src/e/edit/MergeWorkspaceAction.java: ... just like the old Close Workspace...

src/e/edit/EvergreenMenuBar.java: 
src/e/edit/EvergreenTabbedPane.java: ... and is exposed in the same places.

I'd previously suggested that closing a nested workspace could silently merge it.  That would be an expedient way out of the impasse but would be inconsistent.  If the user action is supposed to connote closing all of the files too, then that's what it should always do.  If the user (me) wants to move all the files on a workspace to the next best workspace, then that's something else.

I had been wondering whether the call to moveFilesToBestWorkspaces in closeWorkspace could leave the files where they were if there was no good workspace, so they'd be discarded with the workspace.  I ran into a wall considering what would happen to files that have no good workspace when moveFilesToBestWorkspaces is called after an Edit Workspace.  They shouldn't be closed.  There's no connotation of closing when the workspace root directory is changed even, I would argue, when the file's path is no longer within the workspace.

I think there might be a tasteful way round that wall but I'll check this in anyway as it gets the code back to where no-one should be actively hurting.