src/e/edit/Workspace.java: When asked to reassess the best workspace for a file, don't move it if there's no better choice.  This prevents files with no good home from flying to the selected workspace when that workspaceConfigurationDidChange, regardless of the substance of the change.
src/e/edit/Evergreen.java: Here we retain the previous default choice of the currently selected workspace if there is no better choice but only if the file being opened is not open on any workspace.  This fixes the long standing serious bug whereby you could easily end up with the same file open on two workspaces (if the file didn't belong to either).

Someone else could probably write the array handling here in a less clunky style.  The duplication of setSelectedComponent is mildly distressing too.