src/e/edit/Workspace.java: My fear that r1548's change to EColumn.findWindowByFilename might have violated the now-documented caveat here was groundless...
src/e/edit/Evergreen.java: ... because of a side-effect of the call we use to follow symlinks in Workspace.findIfAlreadyOpen's only caller.  No-one's going to change that but it's worth recording a third reason for using getCanonicalPath, as opposed to, say, checking that the filename refers to a link and then reading the link.  (I realize that wouldn't be possible in Java 5.)  I've put this reason second in the list as it's more important than the seemingly cosmetic reason that follows.

Perhaps the first caveat belongs in EColumn but, if the method there were still called findWindowByTitle, then it clearly wouldn't belong there, as EColumn now does do what it used to say on the tin.