src/e/edit/WorkspaceFileList.java: Following symlinks that lead outside the workspace could lead to our indexer getting stuck in an infinite loop.  Although it was useful to Ed once, it's now getting in Elliott's way, as well you can imagine it might.  Given that we can't remember or devise a use-case where this would be part of the right solution, out it comes, reverting r1060.  r1060 was two days after r1058 broke Ed's link from Evergreen to ../salma-hayek, so he'll presumably spot any remaining need for the r1060 hack quite quickly.
Looking back at r1058, which excludes symlinks within the workspace from the index, I wonder if it wouldn't be more appropriate to include them, providing that the target is a regular file.  As noted earlier today, we canonicalize filenames anyway, so having multiple symlinks to a file won't lead to our opening multiple viewers on the same file (although hard links would).  The motivation for r1058 was a symlink to a directory, which caused a cycle.  Before r1058, the symbolic link test had been in the same place since the initial Subversion import.  (It moved from one file to another and was quite a pain to chase through, so I'm recording that here to save anyone else the bother.)