src/e/edit/RubyDocumentationResearcher.java: Calling a script a binary is confusing, even though bin is an accepted home for many scripts.  The technique already used here to explicitly specify the script interpreter could be used elsewhere to get around the hash-bang problem on Cygwin.
src/e/edit/PythonDocumentationResearcher.java: It's obvious what the interpreter should be here, for example.

src/e/edit/GoToTagAction.java: But the property here, although it's allegedly unused, demonstrates that the hash-bang interpretation really wants to move to ProcessUtilities.  Then part of this change wants reverting, along with at least one previously extant explicit specification of the interpreter.

Although Cygwin bash doesn't like Windows path names, Cygwin Ruby and Python are both happy for their script filename argument to be given in Windows form.  I see that, despite the lack of .rb, evergreen-boilerplate-generator, which is the script that started this discussion, is a Ruby script.