src/e/gui/EMenuBar.java: Refresh every menu item's enabled state before deferring to JMenuBar's processKeyBinding, which searches the menu hierarchy for the first enabled item whose mnemonic matches the keystroke.  This means that Terminator's Find Previous menu item gets enabled without needing the Edit menu to be displayed.  This then activates the Alt-D shortcut, which has been inactive for most Linux (and I think Windows) users since Help, Show Debugging Messages gained a D mnemonic, some months ago.
Alt-D only invokes Find Previous because the Edit menu is searched before the Help menu.  We could nail that loophole shut in GuiUtilities, by having makeMenu always create a sub-class of JMenu that overrides processKeyBinding to return false.  Elliott suggests that, instead, when Terminator is using Alt for shortcuts, we shouldn't apply mnemonics to menus, as this leads to clashes, as with F for File and Find, and bugs, like this.  That sounds good to me.

In making this change, I've hoisted traverseMenu out of our MenuListener, generalized it to collapse some duplication (the reason for which I don't understand) and removed an always-true condition.  I've thrown away our customized versions of JMenuBar.processKeyBinding's recursion and JComponent.processKeyBinding.  The latter had a setEnabled(true) matched by no setEnabled(false), which may warrant revisiting if it turns out to need resurrection.