src/e/edit/Minibuffer.java: Adapt to last night's salma-hayek r2961.
It's a bit difficult to test this, as we still have the relatively recent regression here, about which I'm half-expecting to see some reply, months down the road, when it pops out of Elliott's G1, whereby the text area isn't focused yet, so the Find Next and Find Previous actions aren't yet enabled.

But at least it now compiles.