lib/build/universal.make: Always using cp rather than ln -s is needlessly painful for Cygwin development.  When you click on a file name in Evergreen's errors window, we've previously taken you to the copy under .generated, rather than the one you want to edit.  Fortunately, we made it read-only, to stop accidental edits, but it's still annoying.  It's been more than a year since I added the cygpath code originally and that, it seems, is all we needed to follow Cygwin symlinks in Evergreen.  Thankfully Elliott made me think about what comment the cygpath translation method deserved.
lib/build/per-directory.make: COPY_RULE had two variants, one for use with Windows compilers.  I think this usage is only used with Unix compilers like Cygwin g++.  The other usage is for running WiX, which is a Windows compiler and so sadly still needs the cp variant, at least if we keep the same directory structure (and, in particular, .generated), which I like, not least because it avoids faffing with the version control system to tell it to ignore eg the WiX intermediate files.