src/terminator/view/highlight/SelectionHighlighter.java: I finally got fed up enough of having my clipboard disappear because of new text appearing that I worked out what causes that.  Years, this +-1 bug has been driving me nuts.
Perhaps there's a similar bug in the symmetrical case for lines but I don't understand this whole thing.  How come we allow the selection to be set beyond the bounds of the model but then throw it away, in that case, as soon as we're asked to redoHighlightsFrom?  I see addHighlights disregards its firstLineIndex parameter, meaning that it adds highlights that weren't removed by its caller.