bin/show-alert.rb: All the email address to be customized.  A grep for the callers suggests that most of the call sites don't warrant emails.  Unhandled exceptions do, and not all of our callers have specific user groups to mail, so we still need the old default there.
(This patch best reviewed with white space ignored.)

bin/invoke-java.rb: Allow invoke() to be bypassed in favor of launch().  At the present, Terminator and Evergreen are handling exceptions that occur within their launch0() methods but, once they call invoke(), a nested exception handler is installed which then reports errors from that point forwards.  That nested handler is an unnecessary complexity which tripped me up and which is about to become unfortunate because invoke() doesn't have the email address.