lib/build/filter-build-output.rb: When C++ compilation fails, the intention here was to display the compiler flags.  If it had worked, then I might have worked out the cause of Costantino's compilation woes much earlier.
lib/build/universal.make: His symptoms roughly matched what I see when JDK_ROOT ends up pointing at /usr, which could easily happen if /usr/bin/javac were a regular file, rather than the usual symlink.  (That wasn't the case for Costantino but perhaps a chain of symlinks led to the same result.)  If that happens, then impenetrable compilation errors occur, due to /usr/include/linux being ahead of /usr/include in the search order.  I imagine that jni_md.h won't be in the expected location in this case.  If so, we can use that to catch the problem earlier and report it more clearly.