bin/find-jdk-root.rb: We use find-jdk-root.rb in invoke-java.rb, where we append "/lib/tools.jar".  That was garnering me the following warning in Cygwin-1.7, on the first run of our applications from the command line per Cygwin "boot".  (Cygwin effectively shuts down when the last using application quits.)
cygwin warning:
  MS-DOS style path detected: C:\Program Files\Java\jdk1.6.0_13/lib/tools.jar
  Preferred POSIX equivalent is: /cygdrive/c/Program Files/Java/jdk1.6.0_13/lib/tools.jar
  CYGWIN environment variable option "nodosfilewarning" turns off this warning.
  Consult the user's guide for more details about POSIX paths:
    http://cygwin.com/cygwin-ug-net/using.html#using-pathnames

It wasn't ever our intention to rely on Cygwin working with filenames that are half-Windows, half-Posix style.  This code is used on all platforms, so it's least likely to break if it returns consistently styled results on all platforms.

universal.make translates the returned path to cope with what we were previously returning.  I'd hoped to be able to remove that code in this commit but the space in "Program Files" stymied me.  I had a go at making the makefiles support compilation with a JDK include path containing a space but it was just too hard and already looked too fragile, despite not yet working.

bin/invoke-java.rb: If we can use `cygpath` there, and I think we can, then we can abandon our direct calls into cygwin1.dll here.  That would be a good idea, as they've been deprecated in 1.7, apparently due to containing buffer overflows when used with large paths.  Yes, I know we were being good about asking for the necessary path size but, even so, I think those calls would overflow and they're certainly deprecated and now undocumented.

It seems a shame to see this example code disappear, as it's served us well.  There is a new alternative, but that isn't supported on Cygwin 1.5.  Backquoting cygpath seems to be performing OK even now Evergreen's doing it automatically.

bin/target-os.rb: A Find in Files happened to turn up a related comment which could was due an update.  I could easily be persuaded to always call uname here.