bin/nightly-build.rb: Support a new switch to prevent updating.  Subversion has a not unreasonable, yet unfortunate, habit of stopping older clients from using a work area once it's been used on a newer client.  My "work" area is now shared by Etch and Lenny machines, so needs a Lenny-era svn.  My "nightlies" area is shared by two Etch machines, one of which I don't want to hack with a Lenny svn and one that shares my "work" area with the Lenny box.
bin/ithaki.sh: I need to avoid running update on the old machine, which hasn't the pleasant side-effect of removing two spurious updates from the nightly builds here.  Interestingly, the build still works, even though that looks at the svn version.  Perhaps we cache that in .generated and make can tell that it's up-to-date.