bin/mdorey.sh: Rebuilding the Win32 installation after a disk crash on the nightly build machine, I had to reinvent the exim4 configuration partly from memory (cribbing heavily, as before, from very differently laid-out Debian configurations in the same domain).  Record the incantation here for next time.
This time, I'm trying to build under my home directory...

lib/build/universal.make: ... and, rashly, I thought I'd try to mount my Windows home directory on /home/martind, thus, I hoped, getting round the spaces in "C:\Documents and Settings\martind".  It's fine within the Cygwin world, but we run a couple of native Windows tools, notably javac, candle, light and msiexec.  If this patch is the limit of how much support we'd need for file names containing spaces, then I think we can live with it.  It's when make itself needs to deal with such filenames, in prerequisites and the like, that things really get out of control.