Try harder to use a single VM.  We've talked about making this change on a number of occasions.  Without it, dragging tabs between windows will be arbitrarily constrained by VM boundaries that are otherwise invisible to the user.  Without the ability to drag tabs between windows, it would be less useful to add an option to start new shells in tabs by default, and some users would like that option.  Plus dragging tabs between windows would be cool.  Making this change will make it more likely that this single VM will run out of heap.  I hope that will encourage us to finally do something about that problem.
bin/terminator: Send the whole command line and the directory we were started from.  Uri-encode the arguments so that we can use space to separate them without misinterpreting spaces within arguments.  Reword a very BlueArc-centric comment.

src/terminator/TerminatorServer.java: Apply the reverse transformation to the argument string goedelized in bin/terminator, then hand it off to parseCommandLine.

src/terminator/Terminator.java: Now that bin/terminator is adding a --working-directory to the start of every command line, cope with the case where the user supplies a later --working-directory that's relative to that one.  Perhaps no-one uses --working-directory with a relative path but it was one of the first test cases that sprang to mind.

Remove the code that was allegedly to ignore -xrm argument pairs but which was only ignoring the first argument of the pair.  It was getting in the way of passing a List<String> instead of a String[], not irreparably but, if it ain't working, don't maintain it.  Handle the abnormal --help case with an early return.