src/e/edit/EditServer.java: Looking at this code with a view to sharing it with Terminator, I saw we inflicted a backtrace on the user even for unexpected exceptions that aren't bugs, like this one:
martind@whitewater:~$ edit ~/badger
"~/badger/" is a directory and so cannot be edited with Evergreen.
martind@whitewater:~$ 

Burying them in the log seems (fine by me and) our usual policy.

The exception handler in run() is the one with the non-null ex.getMessage().  The other, around invokeAndWait, is only required to do battle with checked exceptions.  This handler will never actually be called, except in the face of a bug, so logging seems the right choice there too.