bin/terminator: The recent change to use the same VM for more complex command lines is causing grief for BlueArc users of a program called "runmercury".  This kills a program once the terminal returns.  While I could hack on that, it would be more generally useful if TerminatorServer invocations behaved more like independent terminals, as suggested by the code-comment I'm replacing here.  Contrary to that comment, multiple tabs turn out to make no difference...
src/terminator/TerminatorServer.java: ... and we currently only ever create a single TerminatorFrame per command line.  This gives us a window for whose closure we can wait with the new GuiUtilities method.  To wait, in a background thread, for the window to be closed we first need to wait for it to be opened.  To do this, I wanted to use invokeAndWait in place of Terminator.initUi's invokeLater, but invokeAndWait can't be used from the EDT, which is the thread on which the original command line is parsed.  This led me to want an Opener class, like that used by EditServer, which could have two entry points.  That led to the gutting of Terminator.parseCommandLine and its callees.  The differences between this method and the revamped Terminator.parseOriginalCommandLine, then, are:

* TerminatorServer doesn't want to start TerminatorServer
* TerminatorServer must never call System.exit, whereas that is the only way for the EDT to exit with a syntax error on the original command line
* TerminatorServer needs to wait, where it's pointless (if harmless) for the main() thread to wait
* TerminatorServer shouldn't finishGnomeStartup because DESKTOP_STARTUP_ID is wrong

src/terminator/TerminatorOpener.java: Thus the code moved here from Terminator.java no longer calls System.exit when it trips on a syntax error and no longer calls finishGnomeStartup when showing usage in that situation or on receiving --help.  While it's nice that "terminator -e" no longer causes all existing windows to abruptly disappear, it's a bit sad that this, the one piece of syntax that we actually check, wasn't, and still isn't, reflected in the usage.  Oh well.

src/terminator/TerminatorFrame.java: finishGnomeStartup is best delegated to a "finally" clause...

src/terminator/Terminator.java: ... in parseOriginalCommandLine, which will catch all three return paths - syntax requested, syntax error and UI started.  Another rather incidental fix here is that the "arguments" field, which used to be here, is no longer exposed to potential, if unlikely, contention.