The i386 Linux nightly builds have been broken since I once used Lenny (on duezer.us.dev.bluear.com) to access ~/software.jessies.org/work/ and was compelled to upgrade svn on whitewater.us.dev.bluearc.com to match.  I knew I had trouble but thought I'd fixed it all by not calling svn update on ithaki.us.dev.bluearc.com as of "2009-06-26 09:46:22 -0700 mad (3100)".
Since that date, make-version-file.rb has been generating a version number of 0.0.0.  This was mentioned in the cron mails but I hadn't noticed.  The script deliberately doesn't fail (which I would have noticed) in this situation, noting:

  # In a directory not under Subversion control, svnversion(1) says "exported".
  # This happens if you're using Bazaar, say, or nothing, or if you're building from a source tarball.
  # Returning 0 as the version number lets us build without warnings.
  
lib/build/make-version-file.rb: Repurpose some long-obsolete code for parsing "net use" on Cygwin to parse "df" on ithaki and, as in the Cygwin code, run the svnversion command on the file server.