lib/build/make-version-file.rb: In a directory not under Subversion control, svnversion(1) might well say "exported", but a cursory examination of the calling code reveals that we wouldn't run it there.  No-one uses Bazaar and the source tarball doesn't include the .svn directory.  Move the only remaining correct line of the comment to the calling code, where it may be pointless (though we do still offer source tarballs) but is at least harmless.
If we do run svnversion, then, insist that it fails, killing the caller with an informative message and a stack trace otherwise.  This would have prevented the nightly i386 builds from rusting, as they have for the last two or three weeks, due to building a package whose filename includes the version "0.0.0".  After the first day, that filename was always on the server already and hence wasn't overwritten.