lib/build/filter-build-output.rb: Don't wait for the user to confirm that they want to overwrite a read-only file owned by someone else in a writable directory.  They can't see the question that's being asked.
lib/build/universal.make: Use uniform code to move $@.tmp onto $@.  rm -f $@ first, like some of the previous code was doing, to avoid the question entirely.

This stems from my recent Etch to Lenny upgrade, when I did a "sudo make install" to get an up-to-date Terminator on the machine I was doing the upgrade from.  I've been running with the patch for a few days with no problems noticed.